requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
easydict==1.13
flatbuffers==25.2.10
fonttools==4.58.4
grpcio==1.67.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
milvus-lite==2.4.12
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
onnx==1.18.0
onnxruntime-gpu @ file:///home/pandora/Downloads/downloads/Facial-Recognition/onnxruntime_gpu-1.22.0-cp310-cp310-linux_aarch64.whl
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prettytable==3.16.0
protobuf==6.31.1
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pymilvus==2.5.11
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
simsimd==6.4.9
six==1.17.0
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
wcwidth==0.2.13
